+++ release-notes.md
@@ -1,4 +1,5 @@
 Changes since v2.0.0 include:
+- ultrafeeder/rtlsdr: on the fly gain changes without restarting the decoder or container
 - make logs less verbose and improve the diagnostic uploader
 - add ready to go VM images, drop the iob builds, drop the 32bit Pi3 build
 - fix autogain reset for UAT SDRs